+++ simple-chat/src/components/Chat/index.js
@@ -3,5 +3,6 @@ export { default as ChatInput } from './ChatInput';
 export { default as ChatHeader } from './ChatHeader';
 export { default as ControlPanel } from './ControlPanel';
 export { default as MobileControls } from './MobileControls';
+export { default as ScrollToBottomButton } from './ScrollToBottomButton';
 export { default as TipTapEditor } from './TipTapEditor';
 export { default as BubbleMenuContent } from './BubbleMenu';